Rip out old imports module
