+++ tests/TestSuite.hs
@@ -13,6 +13,7 @@ import qualified Language.Haskell.Stylish.Config.Tests
 import qualified Language.Haskell.Stylish.Parse.Tests
 import qualified Language.Haskell.Stylish.Step.Data.Tests
 import qualified Language.Haskell.Stylish.Step.Imports.Tests
+import qualified Language.Haskell.Stylish.Step.Imports.FelixTests
 import qualified Language.Haskell.Stylish.Step.ModuleHeader.Tests
 import qualified Language.Haskell.Stylish.Step.LanguagePragmas.Tests
 import qualified Language.Haskell.Stylish.Step.SimpleAlign.Tests
@@ -21,7 +22,6 @@ import qualified Language.Haskell.Stylish.Step.Tabs.Tests
 import qualified Language.Haskell.Stylish.Step.TrailingWhitespace.Tests
 import qualified Language.Haskell.Stylish.Step.UnicodeSyntax.Tests
 import qualified Language.Haskell.Stylish.Tests
-import qualified Language.Haskell.Stylish.Step.ImportsGHC.Tests
 
 
 --------------------------------------------------------------------------------
@@ -31,7 +31,7 @@ main = defaultMain
     , Language.Haskell.Stylish.Config.Tests.tests
     , Language.Haskell.Stylish.Step.Data.Tests.tests
     , Language.Haskell.Stylish.Step.Imports.Tests.tests
-    , Language.Haskell.Stylish.Step.ImportsGHC.Tests.tests
+    , Language.Haskell.Stylish.Step.Imports.FelixTests.tests
     , Language.Haskell.Stylish.Step.LanguagePragmas.Tests.tests
     , Language.Haskell.Stylish.Step.ModuleHeader.Tests.tests
     , Language.Haskell.Stylish.Step.SimpleAlign.Tests.tests
